feat: ship class and its methods created
Ship class will help at creating each ship type. The ships will have
a length, number of times they were hit and if they sunk or not.
Method hit and isSunk have been created to keep track of the amount of
hits and whether the ship has sunk or not.

Test directory has been created as well for future tests.