feat: Interface created and gameboard rendered
A part of the UI has been created and the gameboard is rendered on the
page where players can click the opponent's board and send attacks.
Logic created for the computer player to make random moves on the board
and announcement of when the game is over.

Small adjustments made to some of the previous methods.